Rename ops.spec to values.spec, fix tests
Update compare tests and move them to values.spec.
